+++ docs/pages/changelog.rst
@@ -12,6 +12,9 @@ Added
 - Added update information on starting screen.
 - Added option to show an overview of all icons (enter Tree.iconOverview() in Console).
 - Added Error count setting for devices to stop communication after multiple consecutive errors.
+- Added option to record videos of plugins.
+- Added option to highlight mouse clicks.
+- Added offline installation instructions.
 
 Fixed
 ~~~~~